# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRISM Engine API",
    "description": "Multi-dimensional STEAM career guidance: psychometric + financial vectorisation, financial constraint solver, parent-student conflict index, market blending, SWOT and roadmaps.\n\nAll /api/v1 responses use the envelope {success, data, error, meta}. MOCK_MODE is currently **True**.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/analytics": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Analytics",
        "operationId": "analytics_api_v1_admin_analytics_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AdminAnalytics_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/data/refresh": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_admin_data_refresh_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RefreshResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/login": {
      "post": {
        "tags": [
          "compat (team aliases)"
        ],
        "summary": "Login",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": [
          "compat (team aliases)"
        ],
        "summary": "Create User",
        "operationId": "create_user_api_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatUserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": [
          "compat (team aliases)"
        ],
        "summary": "Get User",
        "operationId": "get_user_api_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/predict": {
      "post": {
        "tags": [
          "compat (team aliases)"
        ],
        "summary": "Predict",
        "operationId": "predict_api_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatPredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatPredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/results/{result_id}": {
      "get": {
        "tags": [
          "compat (team aliases)"
        ],
        "summary": "Result",
        "operationId": "result_api_results__result_id__get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminAnalytics": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "k_anonymity_threshold": {
            "type": "integer",
            "title": "K Anonymity Threshold",
            "description": "Groups smaller than this are suppressed"
          },
          "students_total": {
            "type": "integer",
            "title": "Students Total"
          },
          "families_linked": {
            "type": "integer",
            "title": "Families Linked"
          },
          "runs_total": {
            "type": "integer",
            "title": "Runs Total"
          },
          "median_run_ms": {
            "type": "number",
            "title": "Median Run Ms"
          },
          "top_recommended_careers": {
            "items": {
              "$ref": "#/components/schemas/CountRow"
            },
            "type": "array",
            "title": "Top Recommended Careers"
          },
          "affordability_class_distribution": {
            "items": {
              "$ref": "#/components/schemas/CountRow"
            },
            "type": "array",
            "title": "Affordability Class Distribution"
          },
          "conflict_band_distribution": {
            "items": {
              "$ref": "#/components/schemas/CountRow"
            },
            "type": "array",
            "title": "Conflict Band Distribution"
          },
          "riasec_code_distribution": {
            "items": {
              "$ref": "#/components/schemas/CountRow"
            },
            "type": "array",
            "title": "Riasec Code Distribution"
          },
          "region_distribution": {
            "items": {
              "$ref": "#/components/schemas/CountRow"
            },
            "type": "array",
            "title": "Region Distribution"
          },
          "suppressed_groups": {
            "type": "integer",
            "title": "Suppressed Groups"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generated_at",
          "k_anonymity_threshold",
          "students_total",
          "families_linked",
          "runs_total",
          "median_run_ms",
          "top_recommended_careers",
          "affordability_class_distribution",
          "conflict_band_distribution",
          "riasec_code_distribution",
          "region_distribution",
          "suppressed_groups"
        ],
        "title": "AdminAnalytics"
      },
...
      "CompatLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "CompatLoginRequest"
      },
      "CompatLoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user_id",
          "role"
        ],
        "title": "CompatLoginResponse"
      },
      "CompatPredictRequest": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "vector": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector",
            "description": "Optional canonical student vector"
          }
        },
        "type": "object",
        "title": "CompatPredictRequest"
      },
      "CompatPredictResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Result id, usable with GET /api/results/{id}"
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "result": {
            "type": "string",
            "title": "Result",
            "description": "Top recommended career name"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "domain_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Domain Scores"
          }
        },
        "type": "object",
        "required": [
          "id",
          "score",
          "result",
          "confidence"
        ],
        "title": "CompatPredictResponse"
      },
      "CompatResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Result id, usable with GET /api/results/{id}"
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "result": {
            "type": "string",
            "title": "Result",
            "description": "Top recommended career name"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "domain_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Domain Scores"
          },
          "top_careers": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "string"
                  }
                ]
              },
              "type": "object"
            },
            "type": "array",
            "title": "Top Careers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "score",
          "result",
          "confidence"
        ],
        "title": "CompatResult"
      },
      "CompatUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role"
        ],
        "title": "CompatUser"
      },
      "CompatUserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "student"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "CompatUserCreate"
      },
...
      "CountRow": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "count"
        ],
        "title": "CountRow"
      },
...
      "DataRefreshRequest": {
        "properties": {
          "datasets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Datasets",
            "examples": [
              [
                "market_signals"
              ]
            ]
          },
          "source": {
            "type": "string",
            "pattern": "^(seed|csv|adapter)$",
            "title": "Source",
            "default": "seed"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DataRefreshRequest"
      },
...
      "Envelope_AdminAnalytics_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminAnalytics"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[AdminAnalytics]"
      },
...
      "Envelope_RefreshResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RefreshResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[RefreshResult]"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "VALIDATION_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "v1/0.1.0"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "description": "true when the payload is a MOCK_MODE fixture",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "Meta"
      },
...
      "RefreshResult": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "completed",
              "failed",
              "dry_run"
            ]
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "counts": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Counts",
            "description": "dataset -> {inserted, updated, skipped}"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "title": "Finished At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "status",
          "source",
          "counts",
          "warnings",
          "started_at",
          "finished_at"
        ],
        "title": "RefreshResult"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "student",
          "parent",
          "educator",
          "admin"
        ],
        "title": "Role"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}